pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hirag-haystack"
version = "0.1.0"
description = "Hierarchical Retrieval-Augmented Generation with Haystack"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["rag", "llm", "haystack", "knowledge-graph", "hierarchical"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "haystack-ai>=2.6",
    "networkx",
    "python-dotenv",
    "python-louvain",
    "tiktoken",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "coverage[toml]",
]
neo4j = [
    "neo4j>=5.0",
]
openai = [
    "openai>=1.0",
    "haystack-ai[openai]",
]
scikit-learn = [
    "scikit-learn",
]
all = [
    "hirag-haystack[dev,neo4j,openai,scikit-learn]",
]

[project.urls]
Homepage = "https://github.com/yourusername/hirag-haystack"
Documentation = "https://github.com/yourusername/hirag-haystack#readme"
Source = "https://github.com/yourusername/hirag-haystack"

[tool.hatch.build.targets.wheel]
packages = ["hirag_haystack"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["hirag_haystack"]

[tool.ruff]
target-version = "py310"
line-length = 100
